+++ CHANGELOG.md
@@ -4,7 +4,7 @@
 
 ## [Unreleased]
 
-###
+### Added
 - Added `pybricks.pupdevices.MarioHub` to control it as a peripheral. It
   cannot be used as a standalone device since it cannot ne updated.
 
@@ -14,9 +14,11 @@
 
 ### Fixed
 - Fixed EV3 light animation stopping when screen is used ([support#2599]).
+- Fixed Powered Up motors not resetting to the absolute value ([support#2620]).
 
 [support#2599]: https://github.com/pybricks/support/issues/2599
 [support#2603]: https://github.com/pybricks/support/issues/2603
+[support#2620]: https://github.com/pybricks/support/issues/2620
 
 ## [4.0.0b7] - 2026-02-19