+++ CHANGELOG.md
@@ -12,6 +12,8 @@
 ### Changed
 - The EV3 Color Sensor now returns `Color.NONE` instead of `None` when no color
   is detected, for consistency with other Color Sensors ([support#2603]).
+- The EV3 Infrared Sensor buttons method now returns a set instead of a list,
+  for consistency with hub buttons.
 
 ### Fixed
 - Fixed EV3 light animation stopping when screen is used ([support#2599]).